fix: rename PIDisag -> PiDisag test files
