--- dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/analytics/event/data/programindicator/disag/PIDisagRecommendedDimensionsTest.java
@@ -1,70 +1,0 @@
-/*
- * Copyright (c) 2004-2022, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- *
- * 3. Neither the name of the copyright holder nor the names of its contributors 
- * may be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.analytics.event.data.programindicator.disag;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import org.hisp.dhis.category.Category;
-import org.hisp.dhis.common.DimensionalObject;
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.springframework.transaction.annotation.Transactional;
-
-/**
- * @author Jim Grace
- */
-@TestInstance(TestInstance.Lifecycle.PER_CLASS)
-@Transactional
-class PIDisagRecommendedDimensionsTest extends AbstractPIDisagTest {
-
-  @Override
-  @BeforeAll
-  protected void setUp() {
-    super.setUp();
-  }
-
-  @Test
-  void testGetRecommendedDimensions() {
-
-    // Given
-    Set<Category> expectedCategories = Set.of(category1, category2, category3);
-
-    // When
-    List<? extends DimensionalObject> recommendations =
-        PiDisagRecommendedDimensions.getRecommendations(eventQueryParams, manager);
-
-    // Then (in any order)
-    assertEquals(expectedCategories, new HashSet<>(recommendations));
-  }
-}
