+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/analytics/event/data/programindicator/disag/PiDisagQueryGeneratorTest.java
@@ -29,10 +29,9 @@
  */
 package org.hisp.dhis.analytics.event.data.programindicator.disag;
 
-import static java.util.Collections.emptyList;
-import static java.util.Collections.emptySet;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import java.util.Date;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
@@ -51,7 +50,7 @@ import org.springframework.transaction.annotation.Transactional;
  */
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 @Transactional
-class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
+class PIDisagQueryGeneratorTest extends AbstractPIDisagTest {
 
   @Autowired private ProgramIndicatorService programIndicatorService;
 
@@ -73,15 +72,20 @@ class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
   void testGetCocSelectColumns_WithoutDimension() {
 
     // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(nonDimensionEventQueryParams);
+    EventQueryParams testParams =
+        new EventQueryParams.Builder()
+            .withProgramIndicator(programIndicator)
+            .withStartDate(new Date())
+            .withEndDate(new Date())
+            .build();
+
+    EventQueryParams params = infoInitializer.getParamsWithDisaggregationInfo(testParams);
 
     Set<String> expectedColumns =
         Set.of(
             "concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end) as \"CategoryId1\"",
             "concat(case when 'filterC' = '' then 'catOption0C' else '' end, case when 'filterD' = '' then 'catOption0D' else '' end) as \"CategoryId2\"",
-            "concat(case when 'filterE' = '' then 'catOption0E' else '' end, case when 'filterF' = '' then 'catOption0F' else '' end) as \"CategoryId3\"",
-            "concat(case when 'filterG' = '' then 'catOption0G' else '' end, case when 'filterH' = '' then 'catOption0H' else '' end) as \"CategoryId4\"");
+            "concat(case when 'filterE' = '' then 'catOption0E' else '' end, case when 'filterF' = '' then 'catOption0F' else '' end) as \"CategoryId3\"");
 
     // When
     List<String> columns = target.getCocSelectColumns(params);
@@ -91,36 +95,26 @@ class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
   }
 
   @Test
-  void testGetCocSelectColumns_WithDimension() {
+  void testGetDisaggregationColumns_WithDimension() {
 
     // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(dataValueSetEventQueryParams);
+    EventQueryParams testParams =
+        new EventQueryParams.Builder()
+            .withProgramIndicator(programIndicator)
+            .addDimension(category2)
+            .withStartDate(new Date())
+            .withEndDate(new Date())
+            .build();
+
+    EventQueryParams params = infoInitializer.getParamsWithDisaggregationInfo(testParams);
 
     Set<String> expectedColumns =
         Set.of(
             "concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end) as \"CategoryId1\"",
-            "concat(case when 'filterE' = '' then 'catOption0E' else '' end, case when 'filterF' = '' then 'catOption0F' else '' end) as \"CategoryId3\"",
-            "concat(case when 'filterG' = '' then 'catOption0G' else '' end, case when 'filterH' = '' then 'catOption0H' else '' end) as \"CategoryId4\"");
+            "concat(case when 'filterE' = '' then 'catOption0E' else '' end, case when 'filterF' = '' then 'catOption0F' else '' end) as \"CategoryId3\"");
 
     // When
     List<String> columns = target.getCocSelectColumns(params);
-
-    // Then (in any order)
-    assertEquals(expectedColumns, new HashSet<>(columns));
-  }
-
-  @Test
-  void testGetCocSelectColumns_WithoutDataValueSet() {
-
-    // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(nonDataValueSetEventQueryParams);
-
-    Set<String> expectedColumns = emptySet();
-
-    // When
-    List<String> columns = target.getCocColumnsForGroupBy(params);
 
     // Then (in any order)
     assertEquals(expectedColumns, new HashSet<>(columns));
@@ -130,15 +124,16 @@ class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
   void testGetDisaggregationGroupByColumns_WithoutDimension() {
 
     // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(nonDimensionEventQueryParams);
+    EventQueryParams testParams =
+        new EventQueryParams.Builder()
+            .withProgramIndicator(programIndicator)
+            .withStartDate(new Date())
+            .withEndDate(new Date())
+            .build();
 
-    Set<String> expectedColumns =
-        Set.of(
-            "concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end)",
-            "concat(case when 'filterC' = '' then 'catOption0C' else '' end, case when 'filterD' = '' then 'catOption0D' else '' end)",
-            "concat(case when 'filterE' = '' then 'catOption0E' else '' end, case when 'filterF' = '' then 'catOption0F' else '' end)",
-            "concat(case when 'filterG' = '' then 'catOption0G' else '' end, case when 'filterH' = '' then 'catOption0H' else '' end)");
+    EventQueryParams params = infoInitializer.getParamsWithDisaggregationInfo(testParams);
+
+    Set<String> expectedColumns = Set.of("\"CategoryId1\"", "\"CategoryId2\"", "\"CategoryId3\"");
 
     // When
     List<String> columns = target.getCocColumnsForGroupBy(params);
@@ -151,30 +146,17 @@ class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
   void testGetDisaggregationGroupByColumns_WithDimension() {
 
     // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(dataValueSetEventQueryParams);
+    EventQueryParams testParams =
+        new EventQueryParams.Builder()
+            .withProgramIndicator(programIndicator)
+            .addDimension(category2)
+            .withStartDate(new Date())
+            .withEndDate(new Date())
+            .build();
 
-    Set<String> expectedColumns =
-        Set.of(
-            "concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end)",
-            "concat(case when 'filterE' = '' then 'catOption0E' else '' end, case when 'filterF' = '' then 'catOption0F' else '' end)",
-            "concat(case when 'filterG' = '' then 'catOption0G' else '' end, case when 'filterH' = '' then 'catOption0H' else '' end)");
+    EventQueryParams params = infoInitializer.getParamsWithDisaggregationInfo(testParams);
 
-    // When
-    List<String> columns = target.getCocColumnsForGroupBy(params);
-
-    // Then (in any order)
-    assertEquals(expectedColumns, new HashSet<>(columns));
-  }
-
-  @Test
-  void testGetDisaggregationGroupByColumns_WithoutDataValueSet() {
-
-    // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(nonDataValueSetEventQueryParams);
-
-    Set<String> expectedColumns = emptySet();
+    Set<String> expectedColumns = Set.of("\"CategoryId1\"", "\"CategoryId3\"");
 
     // When
     List<String> columns = target.getCocColumnsForGroupBy(params);
@@ -187,9 +169,7 @@ class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
   void testGetColumnForSelectOrGroupBy_Select() {
 
     // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(dataValueSetEventQueryParams);
-
+    EventQueryParams params = infoInitializer.getParamsWithDisaggregationInfo(eventQueryParams);
     String expectedColumn =
         "concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end) as \"CategoryId1\"";
 
@@ -204,9 +184,7 @@ class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
   void testGetColumnForSelectOrGroupBy_GroupBy() {
 
     // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(dataValueSetEventQueryParams);
-
+    EventQueryParams params = infoInitializer.getParamsWithDisaggregationInfo(eventQueryParams);
     String expectedColumn =
         "concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end)";
 
@@ -218,47 +196,10 @@ class PiDisagQueryGeneratorTest extends AbstractPiDisagTest {
   }
 
   @Test
-  void testGetCocWhereConditions_withDataValueSet() {
-
-    // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(dataValueSetEventQueryParams);
-
-    List<String> expectedConditions = emptyList();
-
-    // When
-    List<String> conditions = target.getCocWhereConditions(params);
-
-    // Then
-    assertEquals(expectedConditions, conditions);
-  }
-
-  @Test
-  void testGetCocWhereConditions_withoutDataValueSet() {
-
-    // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(nonDataValueSetEventQueryParams);
-
-    Set<String> expectedConditions =
-        Set.of(
-            "length(concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end)) = 11",
-            "length(concat(case when 'filterE' = '' then 'catOption0E' else '' end, case when 'filterF' = '' then 'catOption0F' else '' end)) = 11");
-
-    // When
-    List<String> conditions = target.getCocWhereConditions(params);
-
-    // Then (in any order)
-    assertEquals(expectedConditions, new HashSet<>(conditions));
-  }
-
-  @Test
   void testGetColumnForWhereClause() {
 
     // Given
-    EventQueryParams params =
-        infoInitializer.getParamsWithDisaggregationInfo(dataValueSetEventQueryParams);
-
+    EventQueryParams params = infoInitializer.getParamsWithDisaggregationInfo(eventQueryParams);
     String expectedColumn =
         "concat(case when 'filterA' = '' then 'catOption0A' else '' end, case when 'filterB' = '' then 'catOption0B' else '' end)";
